+++ packages/main-app/.spaas/const/meta.js
@@ -1,3 +1,5 @@
+import {appName} from '../../spaas.config';
+
 const meta = {
   favicon: 'https://bucket-cdn-prd.oss-cn-shenzhen.aliyuncs.com/spaas-favcion/goods.ico', // 需要根据应用来更改对应的图标
 
@@ -23,7 +25,7 @@ const meta = {
   homePageImg: 'https://deepexi.oss-cn-shenzhen.aliyuncs.com/deepexi-services/login-bg.jpg',
 
   // 当前应用名
-  spaName: '应用中心',
+  spaName: appName,
 
   // 菜单资源code
   menuCode: 'spaas-menu',