+++ packages/main-app/src/route.js
@@ -1,13 +1,14 @@
-const path = require('path')
-const resolve = (file)=> {
-  return path.resolve( __dirname, file);
-}
+const path = require('path');
+const resolve = file => {
+  return path.resolve(__dirname, file);
+};
 
-console.log(resolve('./pages/hello.vue'));
-
-export default [
+exports.default = [
   {
     path: '/hello',
-    component: resolve('./pages/hello.vue')
-  }
-]
+    component: resolve('./pages/hello.vue'),
+    title: '欢迎你',
+    appType: '2',
+    enable: false,
+  },
+];