+++ CHANGELOG.md
@@ -1,4 +1,46 @@
-# [](https://github.com/spaasteam/spaas/compare/v1.0.3...v) (2019-11-16)
+# [](https://github.com/spaasteam/spaas/compare/v1.0.6...v) (2019-11-19)
+
+
+
+## [1.0.6](https://github.com/spaasteam/spaas/compare/v1.0.5...v1.0.6) (2019-11-19)
+
+
+### Bug Fixes
+
+* 处理element-ui404的问题 ([dd9d664](https://github.com/spaasteam/spaas/commit/dd9d6647fa43b7331df4e68d15c9e08670126f21))
+
+
+### Features
+
+* 添加spaas nuxt命令 ([77d1a87](https://github.com/spaasteam/spaas/commit/77d1a872ea9345298ce87a4b3cdf5c1f08ed0bad))
+
+
+
+## [1.0.5](https://github.com/spaasteam/spaas/compare/v1.0.4...v1.0.5) (2019-11-18)
+
+
+### Bug Fixes
+
+* 调整发版脚本 ([8f06cfa](https://github.com/spaasteam/spaas/commit/8f06cfa3cf3df66fa231a76eeeb5b99bc43d3fea))
+* 调整自动化脚本执行顺序 ([8935a66](https://github.com/spaasteam/spaas/commit/8935a66f3985fd3b68fd1293f9f0530cc38d9c87))
+
+
+### Features
+
+* 添加github-release-notes依赖 ([52d1f9a](https://github.com/spaasteam/spaas/commit/52d1f9a1865019fefc11a0b55b938741032aa66e))
+* 添加构建发布提醒 ([6d94fc2](https://github.com/spaasteam/spaas/commit/6d94fc214c354116a36ae1408ac33ffe5e57875f))
+* 添加构建发布自动化脚本 ([3487cef](https://github.com/spaasteam/spaas/commit/3487cefad38db75d11acc52bdf28800edf425155))
+* 调整自动化发版脚本 ([fd8d524](https://github.com/spaasteam/spaas/commit/fd8d524958f3e97fa5149a0e9eb25b41e685fbf8))
+
+
+
+## [1.0.4](https://github.com/spaasteam/spaas/compare/v1.0.3...v1.0.4) (2019-11-16)
+
+
+### Features
+
+* 主应用NPM包开发 ([a0b88f8](https://github.com/spaasteam/spaas/commit/a0b88f8626ac3fb454da2affec179d75b6c2ef97))
+* 添加spaas start 启动项目功能 ([f148077](https://github.com/spaasteam/spaas/commit/f148077513769c1894d0adfa067190d698e3cfb3))